separated the code to routes and controllers folder
