+++ backend/controllers/findFoodController.js
@@ -1,0 +1,6 @@
+import { Food } from "../models/Food.js";
+
+export const findFoodController = async (req, res) => {
+	const data = await Food.find({});
+	res.send(JSON.stringify(data));
+};
