+++ backend/routes/food.js
@@ -1,0 +1,8 @@
+import { addFoodController } from "../controllers/addFoodController.js";
+import { findFoodController } from "../controllers/findFoodController.js";
+import { Router } from "express";
+
+export const router = Router();
+
+router.post("/addfood", addFoodController);
+router.get("/food", findFoodController);
